% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fJM.NCC.R
\name{fJM.NCC}
\alias{fJM.NCC}
\title{fJM-NCC: Likelihood inference with full maximum likelihood approach}
\usage{
fJM.NCC(
  obsObject,
  survObject,
  timeVar,
  deltaVar,
  CompRisk = TRUE,
  interFact,
  control = list()
)
}
\arguments{
\item{obsObject}{An \code{lmerMod} or \code{glmerMod} object (longitudinal sub-model).}

\item{survObject}{An object inheriting from class \code{coxph} or class \code{survreg}. In the call to \code{coxph()}
or \code{survreg()}, you need to specify the argument \code{x = TRUE} such that the design matrix is
contained in the object fit. See \bold{Examples}..}

\item{timeVar}{Character string for the time variable in the longitudinal data.}

\item{deltaVar}{Character string for the event indicator column in \code{interFact$data},
0 representing control, k=1, 2... representing competing event k.}

\item{CompRisk}{Logical, whether to include competing risks (default \code{TRUE}).}

\item{interFact}{List with \code{value} (formula) and \code{data} (data frame) for additional covariates or strata.}

\item{control}{A list of control parameters with the following elements:
\describe{
\item{iter.qN}{The number of quasi-Newton iterations. Default is 300.}
\item{tol}{tolerance value for convergence in the log-likelihood.
Default is \code{sqrt(.Machine$double.eps)}.}
\item{parscale}{The \code{parscale} control argument for \code{optim()}.
It should be a numeric vector of length equal to the number of parameters.
Default is 0.01 for all parameters.}
\item{knots}{A numeric vector of the knot positions for the piecewise-constant baseline risk function.}
\item{ObsTimes.knots}{Logical; if \code{TRUE} (default), the positions of the knots are specified
based on the observed event times; otherwise, they are based only on the true event times.}
\item{Q}{The number of internal knots, 4 default.}
\item{GHk}{The number of Gauss–Hermite quadrature points used to approximate the integrals over the random effects.
The default is 15 for one-, two-, or three-dimensional integration when \eqn{N < 2000}, and 9 otherwise,
for the simple Gauss–Hermite rule.}
\item{GKk}{The number of Gauss–Kronrod points used to approximate the integral involved in the calculation
of the survival function. Two options are available: 7 (default) or 15.}
}}
}
\value{
A list of class \code{"JM.NCC"} with estimated coefficients, variance components, Hessian matrix, and other elements.
}
\description{
The fJM-NCC method incorporates all observed data and survival outcome, treating the unobserved
longitudinal outcomes as missing at random. The longitudinal sub-model must be a
generalized linear mixed-effects model (\code{lmerMod} or \code{glmerMod}) and the survival sub-model can
be a Cox model (\code{coxph}) .
}
\details{
Constructs design matrices for longitudinal and survival sub-models, sets up
piecewise-constant baseline hazards, and performs joint likelihood estimation
using Gauss-Hermite and Gauss-Kronrod quadrature.
}
\examples{
\dontrun{
# load example data for linear mixed-effects model
data(obs_full_linear)
data(sur_full_linear)

# Convert survival data to long format for competing risks
full.surdata.long <- JM::crLong(data = sur_full_linear, statusVar="event",
censLevel = 1, nameStrata = "CR", nameStatus = "comp.event")

# Fit Cox model
survObject.full <- survival::coxph(Surv(sur.time, comp.event) ~ gender+strata(CR),
data = full.surdata.long, x = TRUE)

# Fit longitudinal model
obsObject.full <- lme4::lmer(obs.value~sample.time+(1+sample.time|subject),
data=obs_full_linear)

# Fit joint model
res.fjm <- fJM.NCC(obsObject=obsObject.full, survObject=survObject.full,
timeVar = "sample.time", deltaVar="delta", CompRisk=TRUE,
interFact = list(value = ~ CR, data = full.surdata.long), control = list(Q=4))


# load example data for Poisson generalized linear mixed-effects model
data(obs_full_poisson)
data(sur_full_poisson)

obsObject.full <- lme4::glmer(obs.value~sample.time+(1+sample.time|subject),
data=obs_full_poisson, family=poisson(link = "log"))
full.surdata.long <- JM::crLong(data = sur_full_poisson, statusVar="event",
censLevel = 1, nameStrata = "CR", nameStatus = "comp.event")
survObject.full <- survival::coxph(Surv(sur.time, comp.event) ~ gender+strata(CR),
data = full.surdata.long, x = TRUE)
res.fjm <- fJM.NCC(obsObject=obsObject.full, survObject=survObject.full,
timeVar = "sample.time", deltaVar="delta", CompRisk=TRUE,
interFact = list(value = ~ CR, data = full.surdata.long), control = list(Q=4))

}
}
